requirements.txt
# Core dependencies
kiteconnect==4.1.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.2
pytz>=2023.3

# Web dependencies
Flask>=2.3.2
Werkzeug>=2.3.0

# Security
cryptography>=41.0.1
pyOpenSSL>=23.2.0
python-dotenv>=1.0.0

# Visualization and utilities
matplotlib>=3.7.2
tqdm>=4.65.0

# Explicitly add setuptools
setuptools>=65.5.0